server/requirements.txt
aniso8601==9.0.1
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.1
jupyter-core==4.10.0
livereload==2.6.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==304
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tornado==6.1
traitlets==5.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
